# deployment/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: ソーシャルナビゲーターAPI
  description: B2B営業支援プラットフォーム - 企業リスト管理、案件管理、営業ステータス管理を行うシステムのAPI
  version: 1.0.0
  contact:
    name: バジェットアドテクノロジー
servers:
- url: http://localhost:4010
  description: Mock Server
- url: https://api.budget-sales.com/api/v1
  description: Production Server
security:
- bearerAuth: []
...
paths:
...
  /projects/{id}/add-companies:
    post:
      tags:
      - Projects
      summary: 案件に企業を追加
      description: 選択した企業を案件に追加する
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - company_ids
              properties:
                company_ids:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3, 4, 5]
      responses:
        '200':
          description: 追加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  added_count:
                    type: integer
                    example: 5
                  companies:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectCompany'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{id}/available-companies:
    get:
      tags:
      - Projects
      summary: 案件に追加可能な企業一覧
      description: 案件に追加可能な企業を取得。NG判定と追加済み判定を含む
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      - in: query
        name: search
        schema:
          type: string
      - in: query
        name: industry
        schema:
          type: string
      - in: query
        name: page
        schema:
          type: integer
          default: 1
      - in: query
        name: limit
        schema:
          type: integer
          default: 50
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  results:
                    type: array
                    items:
                      allOf:
                      - $ref: '#/components/schemas/CompanyList'
                      - type: object
                        properties:
                          ng_status:
                            type: object
                            properties:
                              is_ng:
                                type: boolean
                              type:
                                type: string
                                enum: [global, client, project]
                                nullable: true
                              reason:
                                type: string
                                nullable: true
                          in_project:
                            type: boolean
                            description: 既に案件に追加済みか
  /projects/{id}/bulk_update_status:
    post:
      tags:
      - Projects
      summary: 案件企業一括ステータス更新
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - company_ids
              - status
              properties:
                company_ids:
                  type: array
                  items:
                    type: integer
                  example:
                  - 1
                  - 2
                  - 3
                status:
                  type: string
                  enum:
                  - 未接触
                  - DM送信予定
                  - DM送信済み
                  - 返信あり
                  - アポ獲得
                  - 成約
                  - NG
                contact_date:
                  type: string
                  format: date
      responses:
        '200':
          description: 一括更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated_count:
                    type: integer
                    example: 3
                  message:
                    type: string
  /projects/{project_id}/ng_companies:
    post:
      tags:
      - Projects
      summary: 案件NG企業設定
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - company_id
              properties:
                company_id:
                  type: integer
                reason:
                  type: string
                  example: クライアントからNG指定
      responses:
        '201':
          description: NG企業設定成功
  /projects/{id}/export_csv:
    get:
      tags:
      - Projects
      summary: 案件CSVエクスポート
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: CSVファイル
          content:
            text/csv:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CompanyList:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 株式会社サンプル
        industry:
          type: string
          example: IT・ソフトウェア
        employee_count:
          type: integer
          example: 150
        revenue:
          type: integer
          example: 500000000
        prefecture:
          type: string
          example: 東京都
        city:
          type: string
          example: 渋谷区
        established_year:
          type: integer
          example: 2010
        website_url:
          type: string
          example: https://example.com
        contact_email:
          type: string
          example: info@example.com
        phone:
          type: string
          example: 03-1234-5678
        notes:
          type: string
          example: 備考
        is_global_ng:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: 2025-01-01 10:00:00+00:00
        updated_at:
          type: string
          format: date-time
          example: 2025-01-15 15:30:00+00:00
        executives:
          type: array
          items:
            $ref: '#/components/schemas/Executive'
...
    Executive:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 山田太郎
        position:
          type: string
          example: 代表取締役
        facebook_url:
          type: string
          example: https://facebook.com/yamada
        other_sns_url:
          type: string
          example: ''
        direct_email:
          type: string
          example: yamada@example.com
        notes:
          type: string
          example: ''
...
    ProjectCompany:
      type: object
      properties:
        id:
          type: integer
          example: 1
        company:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: 株式会社サンプル
            industry:
              type: string
              example: IT・ソフトウェア
        status:
          type: string
          enum:
          - 未接触
          - DM送信予定
          - DM送信済み
          - 返信あり
          - アポ獲得
          - 成約
          - NG
          example: 未接触
        contact_date:
          type: string
          format: date
          nullable: true
          example: null
        notes:
          type: string
          example: ''
        created_at:
          type: string
          format: date-time
          example: 2025-01-10 10:00:00+00:00
        updated_at:
          type: string
          format: date-time
          example: 2025-01-10 10:00:00+00:00
...
  responses:
    BadRequest:
      description: 不正なリクエスト
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: validation_error
              message:
                type: string
                example: 入力データにエラーがあります
              details:
                type: object
                example:
                  name:
                  - この項目は必須です
...